+++ Form/Elements/Youtube.php
@@ -48,4 +48,12 @@ class Youtube extends Element
             }
         });
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getBlockPrefix()
+    {
+        return 'cms_youtube';
+    }
 }