+++ Form/Types/ImageType.php
@@ -27,4 +27,12 @@ class ImageType extends AbstractType
             ])
         ;
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getBlockPrefix()
+    {
+        return 'cms_image';
+    }
 }