+++ Form/Types/ManagedUploadType.php
@@ -122,7 +122,6 @@ class ManagedUploadType extends AbstractType
                 ]);
             }
         });
-
     }
 
     /**
@@ -137,5 +136,13 @@ class ManagedUploadType extends AbstractType
             'attr' => [],
             'upload_dir' => '/uploads/managed-files/',
         ]);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getBlockPrefix()
+    {
+        return 'cms_managed_upload';
     }
 }